% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api.R
\name{neo_tibble}
\alias{neo_tibble}
\title{Turns a list of single column tibbles into a single tibble, "MATCH (s) return s.value, s.value2"}
\usage{
neo_tibble(res)
}
\arguments{
\item{res}{A result from call_neo4j()}
}
\value{
A tibble with
}
\description{
Turns a list of single column tibbles into a single tibble, "MATCH (s) return s.value, s.value2"
}
\examples{
\dontrun{
"MATCH (s) return s.value, s.value2" \%>\% call_neo4j(con) \%>\% neo_tibble()
}
}
